requirements.txt

dj-database-url
Django
django-cors-headers
django-crispy-forms
django-filter
django-import-export
django-redis
django-tailwind
django-widget-tweaks
djangorestframework
djangorestframework_simplejwt
drf-yasg
gunicorn
Jinja2
pillow
prompt_toolkit
psycopg2
psycopg2-binary
Pygments
PyJWT
python-dateutil
python-decouple
python-dotenv
python-slugify
pytz
PyYAML
sqlparse
whitenoise
django-crispy-forms
crispy-bootstrap5
django-allauth
paystackapi
